% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_retention.R
\name{calculate_headcount}
\alias{calculate_headcount}
\title{Calculate headcount}
\usage{
calculate_headcount(df, input_metric, input_group)
}
\arguments{
\item{df}{Tibble containing retention data}

\item{input_metric}{Metric e.g. Fall to Fall}

\item{input_group}{Group e.g. College}
}
\description{
Calculate number of students by cohort, metric, and group.
}
