% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_retention.R
\name{calculate_retention}
\alias{calculate_retention}
\title{Calculate retention}
\usage{
calculate_retention(
  df,
  input_filter_by,
  input_filter_values,
  input_metric,
  input_group,
  input_agency,
  tab
)
}
\arguments{
\item{df}{Tibble containing retention data}

\item{input_filter_by}{Which filters are selected e.g. gender}

\item{input_filter_values}{List of selected values from filters}

\item{input_metric}{Metric e.g. Fall to Fall}

\item{input_group}{Group e.g. College}

\item{input_agency}{Agency e.g. "third week"}

\item{tab}{Which tab? 3 or 4}
}
\description{
Using user-entered filters and columns for grouping,
calculate rentention rates across cohorts.
}
