% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_retention.R
\name{calculate_return_rate}
\alias{calculate_return_rate}
\title{Calculate return rate}
\usage{
calculate_return_rate(df, input_metric)
}
\arguments{
\item{df}{Tibble containing retention data}

\item{input_metric}{Metric e.g. Fall to Fall}
}
\description{
Reshape data and calculate return rate.
}
