% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_2.R
\name{check_update_select}
\alias{check_update_select}
\title{Update drilldown inputs}
\usage{
check_update_select(input, input_a, input_b, rvs)
}
\arguments{
\item{input}{List of all app inputs}

\item{input_a}{Current drilldown input}

\item{input_b}{One of remaining drilldown inputs}

\item{rvs}{Reactive values}
}
\description{
If two drilldown input choices are the same as each other,
replace one with the other
}
