% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_3.R
\name{conditional_filter_input}
\alias{conditional_filter_input}
\title{Create conditional filter input}
\usage{
conditional_filter_input(df, col_name, input_label, id)
}
\arguments{
\item{df}{Tibble containing retentions data}

\item{col_name}{Which column to create filter for}

\item{input_label}{Label for pickerInput}

\item{id}{ID for namespace}
}
\description{
Create conditional panel containing pickerInput
for any groups selected in add_filter.
}
