% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_1.R
\name{create_valuebox}
\alias{create_valuebox}
\title{Create value box}
\usage{
create_valuebox(df, metric_filter, text_color)
}
\arguments{
\item{df}{Tibble containing goal data}

\item{metric_filter}{Which metric to filter data by}

\item{text_color}{Named character vector containing custom colors}
}
\description{
Create shinydashboard valuebox with custom text colours
}
