% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_retention.R
\name{filter_data}
\alias{filter_data}
\title{Filter data}
\usage{
filter_data(df, input_agency, input_filter_by, input_filter_values, tab)
}
\arguments{
\item{df}{Tibble containing retention data}

\item{input_agency}{Agency e.g. "third week"}

\item{input_filter_by}{Which filters are selected e.g. gender}

\item{input_filter_values}{List of selected values from filters}

\item{tab}{Which tab? 3 or 4}
}
\description{
Filter agency and any added group filters.
}
