% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_retention.R
\name{filter_groups}
\alias{filter_groups}
\title{Filter groups}
\usage{
filter_groups(df, filter_by, filter_values, tab)
}
\arguments{
\item{df}{Tibble containing retention data}

\item{filter_by}{Which filters are selected e.g. gender}

\item{filter_values}{List of selected values from filters}

\item{tab}{Which tab? 3 or 4}
}
\description{
Filter values based on user inputs.
}
