% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{get_modal_footer}
\alias{get_modal_footer}
\title{Can you download the data in the modal?}
\usage{
get_modal_footer(retention_version)
}
\arguments{
\item{retention_version}{is the app 'public' or 'private'?}
}
\description{
Determines the modal footer for the sankey pop-up table.
}
