% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{get_raw_retention_data}
\alias{get_raw_retention_data}
\title{Get raw retention data}
\usage{
get_raw_retention_data()
}
\description{
Reads raw retention data from temporary CSV file.
To be replaced by call to database.
}
