% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_retention_data}
\alias{get_retention_data}
\title{Get retention data}
\usage{
get_retention_data()
}
\description{
Reads retention data from RStudio Connect pin.
Requires environment variable RETENTION_VERSION
defined as either "private" or "public".
}
