% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_retention_version}
\alias{get_retention_version}
\title{Get version of retention app}
\usage{
get_retention_version()
}
\description{
Read environment variable which decides
whether to use the public or private version of
the retention data.
}
