% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_summarised_retention_data}
\alias{get_summarised_retention_data}
\title{Get summarised retention data}
\usage{
get_summarised_retention_data()
}
\description{
Reads summarised retention data from temporary CSV file.
To be replaced by call to database.
}
