% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_line_charts.R
\name{goal_line_chart}
\alias{goal_line_chart}
\title{Create tab 1 line chart}
\usage{
goal_line_chart(df, custom_colors)
}
\arguments{
\item{df}{Tibble containing goal rates}

\item{custom_colors}{Named character vector containing custom colors}
}
\description{
Line chart containing return rates for tabs 1
}
