% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_1.R
\name{process_valuebox_data}
\alias{process_valuebox_data}
\title{Process value box data}
\usage{
process_valuebox_data(df, metric_filter, text_color)
}
\arguments{
\item{df}{Tibble containing goal data}

\item{metric_filter}{Which metric to filter data by}

\item{text_color}{Named character vector containing custom colors}
}
\description{
Format goal data for value boxes
  - filter by metric
  - add % suffix to goal value
  - join data with custom colours
}
