% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_line_charts.R
\name{retention_line_chart}
\alias{retention_line_chart}
\title{Create retention line chart}
\usage{
retention_line_chart(df, group)
}
\arguments{
\item{df}{Tibble containing calculated return rate}

\item{group}{What to colour/group line chart by}
}
\description{
Line chart containing calculated return rates for tabs 3 and 4
}
