% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{retentionsApp-package}
\alias{retentionsApp}
\alias{retentionsApp-package}
\title{retentionsApp: Retentions data application}
\description{
Retentions data exploration app-as-a-package.
}
\author{
\strong{Maintainer}: Jumping Rivers \email{info@jumpingrivers.com}

}
\keyword{internal}
